# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MOSTLYVERS API
  version: 1.0.0
  description: Contract for the Android reader and Owner dashboard. FREE_LAUNCH grants non-financial entitlements; payment routes remain reserved and fail closed.
servers:
  - url: http://localhost:5001/v1
  - url: https://api.example.com/v1
...
paths:
...
  /me/push-tokens: { post: { tags: [Reader], operationId: registerPushToken, security: [ { bearerAuth: [] } ], responses: { '204': {description: Registered} } }, delete: { tags: [Reader], operationId: unregisterPushToken, security: [ { bearerAuth: [] } ], responses: { '204': {description: Disabled} } } }
  /me/notifications: { get: { tags: [Reader], operationId: readerNotifications, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Object' } } } }
  /me/notifications/read-all: { post: { tags: [Reader], operationId: readAllNotifications, security: [ { bearerAuth: [] } ], responses: { '204': {description: Read} } } }
  /me/notifications/{notificationId}/read: { parameters: [ {name: notificationId, in: path, required: true, schema: {type: string, format: uuid}} ], post: { tags: [Reader], operationId: readNotification, security: [ { bearerAuth: [] } ], responses: { '204': {description: Read} } } }
...
  /admin/notifications: { get: { tags: [Admin], operationId: adminNotifications, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Page' } } }, post: { tags: [Admin], operationId: sendNotification, security: [ { bearerAuth: [] } ], responses: { '201': { $ref: '#/components/responses/Object' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    internalToken: { type: http, scheme: bearer }
...
  responses:
    Object: { description: Success, content: { application/json: { schema: {type: object, additionalProperties: true} } } }
    Page: { description: Cursor page, content: { application/json: { schema: {$ref: '#/components/schemas/CursorPage'} } } }
...
  schemas:
...
    CursorPage: { type: object, required: [items], properties: {items: {type: array, items: {}}, nextCursor: {type: [string, 'null']}, total: {type: integer}} }
...